+++ src/Atria.Domain/Holders/HolderSnapshotEntry.cs
@@ -4,4 +4,4 @@ namespace Atria.Domain.Holders;
 /// Input line for building a <see cref="HolderSnapshot"/>: one address and the shares it holds at the
 /// cut, plus the resolved investor. The snapshot factory computes each address's share from these.
 /// </summary>
-public readonly record struct HolderSnapshotEntry(string WalletAddress, decimal TokenCount, Guid? InvestorId);
+public readonly record struct HolderSnapshotEntry(string WalletAddress, long TokenCount, Guid? InvestorId);
